Allow cluster to start if we have quorum
We previously waited for all the nodes, but we can instead just wait for
enough nodes that the quorum is the same size.  This means we can start
with 2/3 nodes or 4/5 nodes.

This isn't safe in general, because etcd doesn't have a strong notion of
a quorum, configurable separately from the nodes themselves.