Use unixs for "restore source" endpoint
When restoring a backup for both etcd and etcd-events, I observed a
port-conflict on port 8002/8003.  Switching to unixs:// endpoints (unix
domain sockets) is a little more secure, and avoids the port conflict.

Due to https://github.com/etcd-io/etcd/issues/12450 the syntax is
slightly awkward - we have to make it look like a host and port.