Allow using multiple endpoints for etcd client
The etcdadm init command creates a single member etcd cluster. Other
members can be added to the cluster using the etcdadm join command with
the endpoint of the first member.
If this first member gets deleted, we need to decide on a different endpoint
to be used for the join command for future members. Instead, once the
cluster is created, we should be able to use endpoints of all members with the
join command.
This endpoint only gets used to create an etcd client for listing/adding members.
This commit modifies the ClientForEndpoint method to use all
endpoints that are passed in. This way the etcdadm join command can accept
a comma separated list of etcd members.